Refactored project to use UUID and Money objects:
- Updated Cart and Product models to use UUIDs.
- Implemented Money object for price handling.
- Refactored repository interactions to comply with DDD principles.
- Added and adjusted tests for AddProduct, RemoveProduct, and GetCartQuery handlers.
- Updated Discount Strategies (EveryFifthProductFree, PercentageDiscount) to use Money.
- Ensured all tests reflect accurate business logic and calculations.

+++ src/Domain/Entity/Product.php
@@ -20,7 +20,7 @@ class Product
     private string $name;
 
     #[ORM\Column]
-    private float $price;
+    private int $price;
 
     public function getId(): string
     {
@@ -39,12 +39,12 @@ class Product
         return $this;
     }
 
-    public function getPrice(): float
+    public function getPrice(): int
     {
         return $this->price;
     }
 
-    public function setPrice(float $price): static
+    public function setPrice(int $price): static
     {
         $this->price = $price;
